Extract call tests
